+++ src/pages/testing.tsx
@@ -1,15 +1,9 @@
-import React, {
-  useState,
-  useEffect,
-  useRef,
-  useLayoutEffect
-} from 'react';
-import ColorThemeSwitch from '../components/v2/ColorThemeSwitch';
+import React from "react";
 
 function App() {
-  return (
-      <h1>this up??</h1>
-  )
+    return (
+        <h1>Testing page</h1>
+    )
 }
 
 export default App;